Refactor common code to test.utils
